arch-gcn3,gpu-compute: Initial WG-level barrier impl
Modify S_BARRIER to notify the CU if the barrier has
an immediate or not. If it does, it needs WG-level sync.

Use the dispatcher as the device that handles WG-level sync,
adds in a map of kernels->WGs that is used to store if WGs are running,
at barrier, etc.