gpu-compute: Scoreboard update
Handles WAX dependences in that it stalls because of them.
(However this doesn't fix the initiateAcc issue mentioned in
the previous commit)

Also fixes a bug where instructions will reserve a register for
writing when it won't actually write to the register. This was
an issue because the instruction exec_mask was not written
until the instruction was executed, meaning the exec_mask
was an unknown value when checked in VRF

Change-Id: I5da487dcd0b81e6625b75ed430a6506265612a64